+++ NEWS.md
@@ -1,3 +1,5 @@
+# glymotif (development version)
+
 # glymotif 0.15.0
 
 ## New features