+++ NEWS.md
@@ -8,6 +8,7 @@
 ## New features
 
 * Add `db_motif_info()` to return a tibble of all built-in database motifs, including their names, structures, and alignments. (#35)
+* Expand the built-in motif database to include additional GlycoMotif collections. Now there are 904 unique motifs. (#36)
 
 # glymotif 0.15.0