+++ NEWS.md
@@ -1,3 +1,5 @@
+# glymotif (development version)
+
 # glymotif 0.16.0
 
 ## Breaking changes